+++ benchmarks/platform/cmd/lowlevel/main.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/faustbrian/go-correlation"
-	httpcorrelation "github.com/faustbrian/go-correlation/http"
+	httpcorrelation "github.com/faustbrian/go-correlation/adapters/http"
 	"github.com/faustbrian/go-service"
 	"github.com/faustbrian/go-service/benchmarks/platform/internal/workload"
 	"github.com/faustbrian/go-service/healthhttp"