+++ serverhttp/benchmark_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/faustbrian/go-correlation"
-	httpcorrelation "github.com/faustbrian/go-correlation/http"
+	httpcorrelation "github.com/faustbrian/go-correlation/adapters/http"
 	"github.com/faustbrian/go-service/serverhttp"
 )