+++ serverhttp/example_test.go
@@ -7,7 +7,7 @@ import (
 	"net/http/httptest"
 
 	"github.com/faustbrian/go-correlation"
-	httpcorrelation "github.com/faustbrian/go-correlation/http"
+	httpcorrelation "github.com/faustbrian/go-correlation/adapters/http"
 	"github.com/faustbrian/go-service/serverhttp"
 )